requirements.txt
cffi==1.4.1
characteristic==14.3.0
cryptography==1.1.2
cssselect==0.9.1
decorator==4.0.6
enum34==1.1.1
idna==2.0
ipaddress==1.0.15
ipython==4.0.1
ipython-genutils==0.1.0
lxml==3.5.0
ndg-httpsclient==0.4.0
numpy==1.10.2
pandas==0.17.1
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pyOpenSSL==0.15.1
python-dateutil==2.4.2
pytz==2015.7
queuelib==1.4.2
Scrapy==1.0.3
service-identity==14.0.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
Twisted==15.5.0
w3lib==1.13.0
wheel==0.24.0
zope.interface==4.1.3
